Use the internal URL for nginx proxy pass
`ghost_config_url` contains the external URL for the site. It's used in
the ghost config.js to set the site URL.

By default, the value of `ghost_config_url` is set to the internal ghost
URL (i.e. 127.0.0.1:2368). If the user overrides that setting to set it
to the public domain name for the blog (e.g. http://codebeef.com), while
maintaining the value of `ghost_config_server`, then the blog should be
configures with the correct domain name, and the ghost instance should
still be set to bind to 127.0.0.1:2368, with this value set in the
proxy_pass for nginx.

Currently, changing the `ghost_config_url` also changes the proxy_pass
setting. This commit changes that so that the proxy_pass setting comes
from the `ghost_config_server` values.